Fixed exit(1) if pthreads extension is used.
Thread don't terminate in thread.phpt, if PHP is applied SUHOSIN-patch which
is offered by http://suhosin.org. The reason is that PHP main thread
unexpectedly terminate with exit status 1 where Zend engine is destroying
HashTable.
The SUHOSIN-patched Zend engine is going to check HashTable destructor
whether PHP system is in fine status or not. The pthreads modules overwrite
HashTable destructor when a thread run. The overwriten destructor is not
registered in zend_hash.c. This commit register ovewriting destructor.